day66/dashboard-customization-system/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-socketio==5.11.4
sqlalchemy==2.0.35
asyncpg==0.30.0
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.12
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
bcrypt==4.2.0
alembic==1.13.3
python-dotenv==1.0.0
